iter9 перенес объявление типов
