+++ internal/models/db.go
@@ -9,23 +9,6 @@ import (
 	"strings"
 )
 
-type Event struct {
-	ShortKey    string `json:"shortKey"`
-	OriginalURL string `json:"originalURL"`
-}
-
-type Consumer struct {
-	file *os.File
-	// заменяем Reader на Scanner
-	scanner *bufio.Scanner
-}
-
-type Producer struct {
-	file *os.File
-	// добавляем Writer в Producer
-	writer *bufio.Writer
-}
-
 func NewProducer(filename string) (*Producer, error) {
 	dirPaths := strings.Split(filename, string(filepath.Separator))
